♻️Added Comments to all funcitons in Destination
Added Comments to all funcitons in Destination.tsx.
This commit also contains a different way to display the reviews.
Personally I think this is better because it seperates the info displayed.
Changed the TbarLiftIcon.svg to no longer have a black border.
Changed a letter from capital to lowercase in the review by "John Smith".