♻Refactored filter's chips to use global state
Previously the project used local state for chips which would need to be
lifted up. Now the state can be accesed from anywhere which will make
filtering easier to implement and fill the requirements of using redux.
It may be possible to refactor the chips even more, but will do that in
a later commit.