♻️Restructured the Browse Feature
Browser.tsx now gets the Countries from the backend through a new query
getCountries.
Browsing now sends the user to a specific results pag.
This page contains the destinations in the wanted country.
These destinations are acquired through the query getDestinationsByCountry.
Also added interfaces for the data for the queries.
Fixed linting and prettier errors.