♻️Fixed sorting
Fixed sorting by adding checks if "after" has a value.
If it has a value the backend check whether
the afters X field has a higher value than the X value of the filter.
It then sorts based on the highest value.
Also added the field ElevationDifference to backend.
This reduces the need for calculation in the backend.